extract record-format microcrate from options
