+++ README.md
@@ -167,6 +167,7 @@ The project is organized as a Cargo workspace with the following crates:
 - **copybook-core**: Core parsing and schema types for COBOL copybooks (lexer, parser, AST, layout resolution)
 - **copybook-codec**: Encoding/decoding codecs for COBOL data types and character conversion
 - **copybook-options**: Shared encode/decode option contracts and configuration primitives
+- **copybook-record-format**: Focused fixed-vs-RDW record format contract shared across options/record I/O
 - **copybook-arrow**: Conversion from COBOL layouts/records to Apache Arrow and Parquet
 - **copybook-cli**: Command-line interface with subcommands (parse, inspect, decode, encode, verify, support)
 - **copybook-gen**: Test fixture and synthetic data generation utilities
